requirements.txt
click==7.1.2
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
kafka-python==2.0.2
kazoo==2.5.0
MarkupSafe==1.1.1
psycopg2==2.8.6
pykafka==2.8.0
six==1.16.0
tabulate==0.8.9
Werkzeug==1.0.1